requirements.txt
pandas==2.2.3
transformers==4.46.3
tqdm==4.67.1
pydub==0.25.1
numpy==2.0.2
torch==2.5.1
faster-whisper==1.1.1
rapidfuzz==3.10.1
google-generativeai==0.8.3
typing-extensions==4.12.2
gradio==5.9.1
ipywidgets==8.1.5
